MyHitchhikingSpots for Android
Currently the app allows the user to save a new hitchhiking spot,
update this spot informing if he ‘got a ride’ or ‘took a break’,
evaluate the spot and save a spot for his destination. The app then
lists all the spots saved by the user and the destination spots show a
total of time since the user started hitchhiking until he arrived to
his destination, besides the total rides that he got on his way. The
app currently do not call Google Maps yet to translate the
geo-coordinates of the spots into addresses, but the app is already
done on a way that as soon as those fields are filled they’ll be
automatically displayed where the geo-coordinates are currently
displayed on the list of spots. Also, the current version doesn’t show
any map - the focus here is on tracking the user’s position and giving
options to save ‘hitchhiking spots’ at the chosen positions and add
some extra information. Most of the fields on a spot’s evaluation
screen are the same expected by Hitchwiki when people are adding
hitchhiking spots to their maps.