Added star rating bar for HItchability, stopped getting location updates if user is waiting for a ride, and a couple more small changes
Added Hitchability Rating Bar (with starts) to the Evaluate/Edit form
and to the main view together with the buttons Got a Ride/Take a Break.
Stopped and hid getting updates of user location when user is waiting
for a ride. Fixed the order of the list of spots (now the ordering by
StartDate and Id should work). Changed PT translations Muito ruim/Médio
to Horrível/Normal.