Bug fix and Show user location to the map
Fixed a bug that would cause the app to crash when spotList only had 1
saved spot. Additionally now we show a map for the user current
location on the map as well, so that the user sees where he is and the
map is never empty even if spotList is empty.