Adding constants to manipulate startActivityForResult requests and results
SpotFormActivity should be called by startActivityForResult informing
one of the new request constants added to BaseActivity. In the
onActivityResult event listener the results will be one new result
constants added to that same class (BaseActivity). The main idea here,
besides moving the request values to constants, is to give a result
value that can be used in the future by onActivityResult to update the
spot lists and the UI without needing to reload the whole spot list
from the database as it’s still been done when onResume is called by
MainActivity or MapViewActivity currently.