Refactoring and changing location and map camera listeners
A lot of stuff was changed, but most important is probably the fact
that we stopped our attempts to listen to when the camera was
positioned in a requested position. This really doesn’t seem to be
possible with the current version of Mapbox. The listeners are still
implemented in the code, they’re just not used. We’ll wait for a new
release of Mapbox as it should come with new map listeners that allow
us to now when the map camera was changed programatically or manually
by the user.