Bug fixing: when edit form is opened, the map camera shouldn't be moved to the user's current location
If a spot is being edited, the map camera should be placed at that
spot’s locations. If a new spot is being added, the map camera should
be the same as seen on MyMapsActivity when the user clicked to create
new spot button. We should only move the map camera to the user’s
current location if they request to do so (by clicking the Locate
button).