Add fragment to back stack only after the first fragment has been opened
Adding a transaction to the back stack, guarantees that when user
clicks on the Back button they're sent back to the fragment they're
come from.
If it's the first fragment being loaded, we want the app to close when
user clicks on the Back button, therefore we don't want to add the
fragment to the back stack.