Replacing SettingsFragment for ToolsActivity
ToolsActivity will update the list of spots through the import option,
therefore we’ll keep it as a separate activity. In this way, once the
user is done importing a new list of spots and he navigates back, the
list of spots gets automatically reloaded without the need of adding
any extra logic.