Bug fix - Avoid keeping HW spots when navigating out of HW Maps
If the list of spots was too big, a fatal crash was happening.

How:
Opened HW Maps after downloading loads of spots. Spots were shown on
map as expected. Clicked to open one of the spots, then the next view
opened but it soon crashed with exception
android.os.TransactionTooLargeException