Bug fixing - making sure that the coordinate of spots are properly compared
A lot of refactoring have been made but the most relavant change on this commit is about comparing spots being imported with existing spots.
It is possible that the coordinate of spots being imported are the same coordinate of a existing spot but if the coordinate of the spots being imported have been converted into float then they won't be exactly the same. Ex: If the latitude -43.934732423 was converted into -43.934 while exporting, then our condition here will be "latitude BETWEEN -43.934 AND -44.934".