Bug fixing - export geocoordinate without lossing precision data
Up to now the precision of the geocoordinate of the spots were getting lost when user exported their database. This was happening because cursor.getString(i) converts Double values into Float values. If the latitude was -43.934732423 then after converted it was being saved as -43.934. With this commit latitude should be saved as is, -43.934732423.