V2.4.3 (#94)
* Refactoring - apply automatic formating document - nothing relevant changed

* Bug fix - avoiding reference to null objects exceptions

* Update to version 36 (2.4.2)

* Bug fix - Avoid null reference exception

* Refactoring - reducinh the number of repeated string resources

* Updating gradle version to 3.5.2

* Bug fixing - making sure that the coordinate of spots are properly compared

A lot of refactoring have been made but the most relavant change on this commit is about comparing spots being imported with existing spots.
It is possible that the coordinate of spots being imported are the same coordinate of a existing spot but if the coordinate of the spots being imported have been converted into float then they won't be exactly the same. Ex: If the latitude -43.934732423 was converted into -43.934 while exporting, then our condition here will be "latitude BETWEEN -43.934 AND -44.934".

* Bug fixing - export geocoordinate without lossing precision data

Up to now the precision of the geocoordinate of the spots were getting lost when user exported their database. This was happening because cursor.getString(i) converts Double values into Float values. If the latitude was -43.934732423 then after converted it was being saved as -43.934. With this commit latitude should be saved as is, -43.934732423.

* Removing unused imports

* Bug fixing - Trying to prevent crashes when screen rotates

* Prevent crash when activity is null

* Replacing progress message while drawing maps

* Update app version to 37 (2.4.3)